# src/AppForSEII2526.Web/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
    "/api/Compras/GetDetallesCompra": {
      "get": {
        "tags": [
          "Compras"
        ],
        "operationId": "GetDetallesCompra",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetalleDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetalleDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetalleDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Compras/CrearCompra": {
      "post": {
        "tags": [
          "Compras"
        ],
        "operationId": "CrearCompra",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearCompraDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearCompraDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CrearCompraDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetalleDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetalleDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetalleDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Ofertas/GetDetallesOferta": {
      "get": {
        "tags": [
          "Ofertas"
        ],
        "operationId": "GetDetallesOferta",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaDetalleDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaDetalleDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaDetalleDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Ofertas": {
      "post": {
        "tags": [
          "Ofertas"
        ],
        "operationId": "CrearOferta",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearOfertaDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearOfertaDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CrearOfertaDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaDetalleDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaDetalleDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfertaDetalleDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CompraDetalleDTO": {
        "required": [
          "apellidosCliente",
          "direccionEnvio",
          "fechaCompra",
          "nombreCliente"
        ],
        "type": "object",
        "properties": {
          "nombreCliente": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "apellidosCliente": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "direccionEnvio": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "fechaCompra": {
            "type": "string",
            "format": "date"
          },
          "precioTotal": {
            "type": "number",
            "format": "currency",
            "nullable": true
          },
          "herramientasCompradas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompraItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CompraItemDTO": {
        "required": [
          "cantidad",
          "nombre"
        ],
        "type": "object",
        "properties": {
          "herramientaid": {
            "type": "integer",
            "format": "int32"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string"
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "precio": {
            "type": "number",
            "format": "currency"
          }
        },
        "additionalProperties": false
      },
      "CrearCompraDTO": {
        "required": [
          "correoElectronico"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "apellido": {
            "type": "string",
            "nullable": true
          },
          "correoElectronico": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "numTelefono": {
            "maxLength": 50,
            "minLength": 1,
            "pattern": "^\\+?\\d{10,15}$",
            "type": "string",
            "nullable": true
          },
          "direccionEnvio": {
            "type": "string",
            "nullable": true
          },
          "metodoPago": {
            "$ref": "#/components/schemas/formaPago"
          },
          "herramientasCompradas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompraItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CrearOfertaDTO": {
        "type": "object",
        "properties": {
          "fechaInicio": {
            "type": "string",
            "format": "date"
          },
          "fechaFin": {
            "type": "string",
            "format": "date"
          },
          "porcentaje": {
            "maximum": 100,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "tiposMetodoPago": {
            "$ref": "#/components/schemas/tiposMetodoPago"
          },
          "tiposDirigidaOferta": {
            "$ref": "#/components/schemas/tiposDirigidaOferta"
          },
          "ofertaItem": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OfertaItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "OfertaDetalleDTO": {
        "required": [
          "fechaFin",
          "fechaInicio",
          "fechaOferta"
        ],
        "type": "object",
        "properties": {
          "Id": {
            "type": "integer",
            "format": "int32"
          },
          "fechaInicio": {
            "type": "string",
            "format": "date"
          },
          "fechaFin": {
            "type": "string",
            "format": "date"
          },
          "fechaOferta": {
            "type": "string",
            "format": "date"
          },
          "nombreUsuario": {
            "type": "string",
            "nullable": true
          },
          "metodoPago": {
            "$ref": "#/components/schemas/tiposMetodoPago"
          },
          "tiposDirigidaOferta": {
            "$ref": "#/components/schemas/tiposDirigidaOferta"
          },
          "herramientasAOfertar": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OfertaItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OfertaItemDTO": {
        "type": "object",
        "properties": {
          "herramientaId": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "material": {
            "type": "string",
            "nullable": true
          },
          "fabricante": {
            "type": "string",
            "nullable": true
          },
          "precio": {
            "type": "number",
            "format": "float"
          },
          "precioOferta": {
            "type": "number",
            "format": "float"
          },
          "porcentaje": {
            "maximum": 100,
            "minimum": 0,
            "type": "integer",
            "format": "currency"
          }
        },
        "additionalProperties": false
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "formaPago": {
        "enum": [
          "Efectivo",
          "TarjetaCredito",
          "PayPal"
        ],
        "type": "string"
      },
...
      "tiposDirigidaOferta": {
        "enum": [
          "Socios",
          "Clientes"
        ],
        "type": "string"
      },
      "tiposMetodoPago": {
        "enum": [
          "Tarjeta",
          "PayPal",
          "Efectivo"
        ],
        "type": "string"
      }
    }
  }
}